Added the EXECUTE priority.
Under this priority, an event is listened to for cancelling it and
overriding its execution.

This priority is useful for plugins that do not change how the incident
occurs, i.e. semantically the incident represented by the event still
occurs, but it is technically cancelled to prevent the event from being
xecuted by the calling context.