PluginManager: Skip methods not declared by instanceof Listener when registering handlers This is quite an interesting bug. If you have
class A{
    public function onMove(PlayerMoveEvent \$event){} //shouldn't be a handler because this class isn't a Listener
}

class B extends A{}

registerEvents(new B, \$plugin);

then A::onMove() will be registered as an event handler even though A is not an instanceof Listener.

This was observed by noting that plugins which do something like "extends PluginBase implements Listener" causes registerEvents() to try and register PluginBase methods as event handlers, which could lead to astonishing behaviour.