Player: Separate disconnection logic into disconnect() API method close() should NOT be used by plugins anymore.
disconnect() will do the logic that won't cause things to crash. close() will handle breaking cyclic references and cleanup only.

This fixes a multitude of bugs relating to player kicking and closing during events. This fixes #1239 and a range of other related (but not reported on the issue tracker) bugs.