Added experimental support for ZOPFLI chunk compression ZOPFLI can produce outputs 10-20% smaller than ZLIB level 9 does, but at a huge extra CPU cost.
This huge cost (generally a couple of orders of magnitude) may be worth it for big servers which get lots of traffic, whose maps are static (resulting in a very high percentage of cache hits).
If the map is immutable, the tradeoff between CPU cost and bandwidth is massively skewed in favour of saving bandwidth, since the cache can be primed once and then used many times.