Fix ServerSettings client-side bug with hack.
In order to use ServerSettings normally, it is necessary to send the response with a delay.
This is very weird, and if you send a response at the same time as the request, it may or may not work.
It turns out that it can be avoided by sending after a few dozen ticks.
However, I could not use TaskScheduler in Player, so I had to use EntityBaseTick.