Separated TimingsRecord from TimingsHandler (different lifetimes) this fixes #3906. TimingsHandler will now be automatically destroyed when there are no TimingsRecord referencing it and its owner has no references left to it. TimingsRecord will be automatically destroyed at the end of the timings session (if active).
This fixes records from temporary timers being lost after the timings handlers are removed, while also fixing TimingsHandlers piling up regardless of whether timings are running or not.
This opens the doors to have more volatile timers, such as per-player/per-session timers, which are destroyed on player quit, but which won't result in the timings disappearing from the final timings result.