Add a way to explicitly blame a plugin for crashing the server We often see cryptic crashes like these: https://crash.pmmp.io/view/5617460
These are actually caused by bogus plugin code, but due to a PHP bug, they are reported in PluginManager instead of at their actual location.
This is not likely to be addressed in the near future (it's been a bug for a very long time), and in the meantime it's adding pollution to the crash archive.

A new crashdump plugin_involvement 'blame' status is added, which the crash archive will need to support. I think for now it's important to distinguish between these at the CA level, so that any bugs in the call proxy can be identified correctly.