Stop throwing exceptions on extra properties.
This code currently throws errors when properties other than the base required ones are added. This can be from resource packs created by "bridge.", where the IDE adds a "generated_with" property to the manifest. Ultimately, this leads to PM flagging it as an undefined property. Should this be set true again, it is recommended that extra work be put into ensuring that the Undefined Property flag doesn't detect extra defined properties.