Throw exception if an event handler is a generator
Plugins that copy code from await-generator examples may forget to wrap their code in `Await::f2c`, and end up having a function that merely returns a `Generator`, executing nothing semantically without triggering any error.

It would be nicer to require them to declare `: void`, but that might break a lot of plugins. Meanwhile, handler functions that are actually generator functions are doubtlessly bugs.

By the way, `ReflectionFunctionAbstract::isGenerator()` only reports functions with a literal `yield`, and does not report functions that declare `Generator` as a return type (although that is equally pointless in the case of event handlers).