TextFormat: Introduced rebase() static method
this function changes the "base" format of a string, i.e. the colours and formats that are used after a RESET (&r) formatting code.

An example of where this is needed is in the logger.

The following code:
$logger->notice("I'm a " . TextFormat::RED . "special" . TextFormat::RESET . " cookie");

causes the "cookie" part of the message to show as grey, instead of the expected aqua for NOTICE level messages.

This function operates by directly suffixing the base format to any &r formatting code found.