Micro optimization gamemode changer
During the call of the `internalSetGameMode` method, the `setFlying` and `setHasBlockCollision` methods are called and if the value passed in their parameters is different from the current one, they call the `syncAbilities` method of the NetworkSession class, but after the execution of the `internalSetGameMode` method, the `setGamemode` method performs a validation that ends up calling the `despawnFromAll` and `spawnToAll` methods without considering if the player's previous gamemode was the spectator method. At the end of this whole process the `syncGameMode` method is called which also calls the `syncAbilities` method, so we can conclude that there is no need for the 2 calls of the `syncAbilities` method previously.