A small optimisation on gamemode change
In the method `internalSetGameMode´ is performed 2 calls of the method `syncAbilities´ without any need, since the method `syncGamemode´ performs the method call at the end of the whole process. The validation that is in the `setGamemode´ method does not consider the previous gamemode of the player. This way it validates if the player is in spectator mode and calls the `despawnFromAll´ method otherwise it just calls the `spawnToAll´ method even if the player's gamemode was not Spectator before.