Support dynamically describing arbitrary enums Previously, we were using codegen to support describing a fixed set of enums.
Instead, we implement an enum() function, allowing any enum implementing RuntimeDataEnum to be described.
This enum can be implemented by EnumTrait users with no additional work required, and it can also be implemented by native enums by making getAll() return cases().

This implementation is:
- faster
- requires way less code
- does not require a build step
- way more flexible

This fixes #5877, increasing the range of stuff that plugins are now able to do.