Fixed self updates not working correctly
when a self update occurs, it's usually because the block itself has just been set. This might be because of growth, placement, or a variety of other factors.
in such cases, all surrounding checks must be done to ensure correct behaviour.
in essence, a block that was just created has had all of its neighbours changed at once, so this is kinda elegant.
this also removes the need for a dedicated self flag, which was only checked when needing to workaround this bug anyway.