Mace Implementation
JavierLeon9966 commented 2 hours ago
Smash particles that increase per distance fallen
Damage to the entity increased per distance fallen
Knock back to nearby entities to the smashed entity
The smash sounds shouldn't be played when breaking a block or normally hitting an entity. For example, the heavy smash is played when falling too high
The player falling motion is reset when smashing something
You're using applyDamage wrong. That's for changing the item's durability