Second look at extracting generator executors from World this is a greatly simplified alternative to #6668.
I realized we probably don't actually want to let the executor decide how promises
are managed, since the system is built to be async anyway.

This new iteration makes it far less complex to implement a custom executor, without
needing to concern oneself with complex promise management. This means it should be
fairly practical to make custom executors with this.