Partial support for latest movement settings MS is due to remove the non-server-auth versions of all of this stuff.
Fortunately v3 server auth movement works just fine without any changes,
although we will need to start sending player tick in some packets if
someone wants to actually use the rewind stuff.

Server auth block breaking is "working" but not done yet - I need to
figure out how the MineBlock action is supposed to be processed. However
this version at least allows the actions to work even if it is noisy on
the network for now.