Alternative solution for infinite effects
The problem is that we need to ensure an updating ticker for periodic effects to continue
working correctly, but those are typically implemented using the duration ticker.

We don't want to BC break anyone implementing custom effects in PM5, so this is the best
we can do for now.

I did consider a bitflag on the duration, but that would've been subjected to math
operations like the ones done by SplashPotion and AreaEffectCloud, which would've produced
bizarre results and potentially wiped out the infinite status of the effect. To avoid
complicating setDuration(), a separate property seemed like the next best idea.

In this implementation, durations will underflow back to Limits::INT32_MAX when they
would have expired, ensuring a continuous counter. This might cause unexpected behaviour
for consumers of the API, but I don't see another viable solution for PM5.