Skip decoding packets which are never handled by the current PacketHandler
Using reflection, we can determine which packets a PacketHandler class will
actually handle and skip decoding the ones that it won't handle. This is
a significant upgrade to server security and possibly performance.

The reflection metadata is calculated the first time a handler class is
used and cached afterwards. This means that the performance impact of
this sorta-hack is negligible.

Previously, any designated serverbound packet would be decoded by the
server at any time, regardless of whether it would actually be handled,
which allowed attackers to waste CPU time by making the server chew
through packets that it wouldn't actually do anything with anyway. While
this doesn't fix any outstanding security vulnerability, it does
significantly limit what attackers can do and when.

This change will likely affect plugins using DataPacketReceiveEvent
implement features not supported by PocketMine-MP. If a plugin needs to
handle a packet that isn't handled by PocketMine-MP's core packet
handlers, it can declare an event handler for DataPacketDecodeEvent with
@handleCancelled in the doc comment, and use ev->uncancel() to force
packets to continue to decoding. It's best to filter by packet ID when
doing this, and only uncancel when it sees a packet it wants to handle,
to avoid negating the benefits of this change.